pyproject.toml
[project]
name = "zerobounce-mcp"
version = "0.1.0"
description = "A Model Context Protocol (MCP) server for the ZeroBounce email validation service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.0",
    "asyncio>=3.4.3",
    "aiofiles>=0.8.0",
    "zerobouncesdk>=1.1.1",
]
